feat: users can now search for a mentor's name
