+++ src/components/MainMentorList.js
@@ -1,64 +1,76 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom'
-import * as Api from '../utils/Api'
+
+const Tags = (props) => {
+  if (props.content === "" || props.content === "[]") {
+    return null
+  } else {
+    let tags = JSON.parse(props.content)
+    // We want to keep tags to a max of 3
+    // ABaixo de 1290 pomos so 2 + o simbolo
+    if (tags.length > 3) {
+      if (props.small) {
+        tags = tags.slice(0, 2)
+      } else {
+        tags = tags.slice(0, 3)
+      }
+      tags.push({
+        text: "+"
+      })
+    }
+    if (tags.length === 3 & props.small) {
+      tags = tags.slice(0, 2)
+      tags.push({
+        text: "+"
+      })
+    }
+    let hello = tags.map((tag, i) =>
+      <li key={i} className='flex align-items-center mentor-tag'>{tag.text}</li>
+    )
+    return hello
+  }
+}
 
 export default class MainMentorList extends Component {
 
-  constructor (props) {
-    super(props)
-    this.state = {
-      mentors: []
+  render() {
+    let small = false
+    if (document.body.clientWidth < 1290) {
+      small = true
     }
-  }
-
-  componentDidMount () {
-    Api.getRandomMentors().then((res) => {
-      this.setState({
-        mentors: res.mentor ? res.mentor : []
-      })
-    })
-  }
-
-  mentorTagsToElement = (tags) => {
-    return tags.map((tag) => {
+    if (this.props.mentors !== [] && this.props.mentors !== undefined) {
       return (
-        <li className='mentor-tags-list-element'>{tag}</li>
-      )
-    })
-  }
-
-  mentorToElement = (mentor) => {
-    return (
-      <Link to={'/people/' + mentor.keycode}>
-        <div className='flex'>
-          <img className='mentor-card-image' src={mentor.profilePic} alt={mentor.name}/>
-          <div className='mentor-card-info'>
-            <h1>{mentor.name}</h1>
-            <p>{mentor.role} @ <Link to={'/companies/' + mentor.company}>{mentor.company}</Link></p>
-            <p>{mentor.bio}</p>
+        <main id='home'>
+          <div className="container grid">
+            <div className="mentor-list">
+              {this.props.mentors.map((mentor) => {
+                return (
+                  <Link to={mentor.keycode}>
+                    <div className='card mentor-card flex justify-center'>
+                      <div>
+                        <img className='mentor-profilepic' src={mentor.profilePic} alt={mentor.name} />
+                        <div className='mentor-info'>
+                          <h1 id='name' className='font-150 font-weight-normal'>{mentor.name}</h1>
+                          <p id='role-company'>{mentor.role} at {mentor.company}</p>
+                          <p id='bio' style={{ WebkitBoxOrient: 'vertical'}}>{mentor.bio}</p>
+                          <ul id='tags' className='flex'>
+                            <Tags content={mentor.tags} small={small} />
+                          </ul>
+                        </div>
+                      </div>
+                    </div>
+                  </Link>
+                )
+              })}
+            </div>
           </div>
-          {/*DEBUG for now <ul className='mentor-card-tags'>
-            {this.mentorTagsToElement(mentor.tags)}
-          </ul> */}
-        </div>
-      </Link>
-    )
-  }
-
-  render () { // For each mentor...
-    if (this.state.mentors !== []) {
-      return (
-        <div className="mentor-list">
-          {this.state.mentors.map((mentor) => {
-            return this.mentorToElement(mentor)
-          })}
-        </div>
+        </main>
       )
     } else {
       return (
-        <div>
-          Loading...
-        </div>
+        <main id='home'>
+          No mentors found
+        </main>
       )
     }
   }
